pyproject.toml
[tool.poetry]
name = "udacity-ia-scholarship"
version = "0.1.0"
description = "For the udacity AI scholarship"
authors = ["dpalmasan <dpalmasan@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
Pillow = "^9.1.0"
video-indexer = "^0.1.8"
azure-cognitiveservices-vision-face = "^0.6.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | grpc_api
)/
'''